Convert _read_sockaddr to JS-only function
This function returns as JS object and is not designed to be
called directlys from native code.

As such is it also meaningless to include it in deps_info.py.
This is because only symbols contains in native object files
can appear in deps_info.py.  See #13431 for more on this.